rename sourcode dir to shortlister
