+++ shortlister/controller.py
@@ -1,6 +1,6 @@
-from source_code.view import View
+from shortlister.view import View
 from readchar import readkey
-from source_code.model import load_shortlist,save_shortlist
+from shortlister.model import load_shortlist,save_shortlist
 from startfile import startfile
 
 class Controller:
