+++ shortlister/model.py
