+++ shortlister/view.py
