+++ tests/test.py
@@ -1,4 +1,4 @@
-from source_code.model import Applicant, Criterion, Role, Shortlist
+from shortlister.model import Applicant, Criterion, Role, Shortlist
 import csv
 from pathlib import Path
